refactor: 미사용 dto 제거 및 Line 관련 dto 이름 변경
