+++ src/main/java/wooteco/subway/dto/LineRequest.java
@@ -4,17 +4,38 @@ public class LineRequest {
 
     private String name;
     private String color;
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
 
-    public LineRequest(String name, String color) {
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
         this.name = name;
         this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
     }
 
     public String getColor() {
         return color;
     }
 
-    public String getName() {
-        return name;
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
     }
 }
