+++ src/main/java/wooteco/subway/dto/LineResponse.java
@@ -1,15 +1,19 @@
 package wooteco.subway.dto;
 
+import java.util.List;
+
 public class LineResponse {
 
     private Long id;
     private String name;
     private String color;
+    private List<StationResponse> stations;
 
-    public LineResponse(Long id, String name, String color) {
+    public LineResponse(Long id, String name, String color, List<StationResponse> stations) {
         this.id = id;
         this.name = name;
         this.color = color;
+        this.stations = stations;
     }
 
     public Long getId() {
@@ -22,5 +26,9 @@ public class LineResponse {
 
     public String getColor() {
         return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
     }
 }
