--- src/main/java/wooteco/subway/dto/LineWithStationResponse.java
@@ -1,34 +1,0 @@
-package wooteco.subway.dto;
-
-import java.util.List;
-
-public class LineWithStationResponse {
-
-    private Long id;
-    private String name;
-    private String color;
-    private List<StationResponse> stations;
-
-    public LineWithStationResponse(Long id, String name, String color, List<StationResponse> stations) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.stations = stations;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-}
