+++ src/main/java/wooteco/subway/ui/LineController.java
@@ -2,8 +2,8 @@ package wooteco.subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.dto.LineWithStationRequest;
-import wooteco.subway.dto.LineWithStationResponse;
+import wooteco.subway.dto.LineRequest;
+import wooteco.subway.dto.LineResponse;
 import wooteco.subway.service.LineService;
 
 import java.net.URI;
@@ -20,25 +20,25 @@ public class LineController {
     }
 
     @PostMapping
-    public ResponseEntity<LineWithStationResponse> createLine(@RequestBody final LineWithStationRequest lineRequest) {
-        final LineWithStationResponse lineResponse = lineService.save(lineRequest);
+    public ResponseEntity<LineResponse> createLine(@RequestBody final LineRequest lineRequest) {
+        final LineResponse lineResponse = lineService.save(lineRequest);
         return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
     }
 
     @GetMapping
-    public ResponseEntity<List<LineWithStationResponse>> getLines() {
-        final List<LineWithStationResponse> lineResponses = lineService.findAll();
+    public ResponseEntity<List<LineResponse>> getLines() {
+        final List<LineResponse> lineResponses = lineService.findAll();
         return ResponseEntity.ok().body(lineResponses);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<LineWithStationResponse> getLine(@PathVariable final Long id) {
-        final LineWithStationResponse lineResponse = lineService.findById(id);
+    public ResponseEntity<LineResponse> getLine(@PathVariable final Long id) {
+        final LineResponse lineResponse = lineService.findById(id);
         return ResponseEntity.ok().body(lineResponse);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> updateLine(@PathVariable final Long id, @RequestBody final LineWithStationRequest lineRequest) {
+    public ResponseEntity<Void> updateLine(@PathVariable final Long id, @RequestBody final LineRequest lineRequest) {
         lineService.update(id, lineRequest);
         return ResponseEntity.ok().build();
     }
